Actualizacion de repositorio para despliegue automatico

+++ src/main/java/com/java/demo/ecommerceapi/controller/BrandController.java
@@ -8,9 +8,12 @@ import org.springframework.web.bind.annotation.*;
 import com.java.demo.ecommerceapi.model.Brand;
 import com.java.demo.ecommerceapi.service.BrandService;
 
+import io.swagger.v3.oas.annotations.tags.Tag;
+
 import java.util.List;
 import java.util.Optional;
 
+@Tag(name = "Authentication", description = "Endpoints for brand management")
 @RestController
 @RequestMapping("/api/brands")
 public class BrandController {
@@ -24,14 +27,18 @@ public class BrandController {
     @GetMapping
     public ResponseEntity<List<Brand>> getAllBrands() {
         List<Brand> brands = brandService.getAllBrands();
-        return new ResponseEntity<>(brands, HttpStatus.OK);
+        if(brands.isEmpty()){
+            return new ResponseEntity<>( HttpStatus.NOT_FOUND);
+        }else{
+            return new ResponseEntity<>(brands, HttpStatus.OK);
+        }
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Optional<Brand>> getBrandById(@PathVariable Long id) {
+    public ResponseEntity<Brand> getBrandById(@PathVariable Long id) {
         Optional<Brand> brand = brandService.getBrandById(id);
         if (brand.isPresent()) {
-            return new ResponseEntity<>(brand, HttpStatus.OK);
+            return new ResponseEntity<>(brand.get(), HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -40,7 +47,11 @@ public class BrandController {
     @PostMapping
     public ResponseEntity<Brand> createBrand(@RequestBody Brand brand) {
         Brand createdBrand = brandService.createBrand(brand);
-        return new ResponseEntity<>(createdBrand, HttpStatus.CREATED);
+        if(createdBrand !=null){
+            return new ResponseEntity<>(createdBrand, HttpStatus.CREATED);
+        }else{
+            return new ResponseEntity<>( HttpStatus.BAD_REQUEST);
+        }
     }
 
     @PutMapping("/{id}")
@@ -55,7 +66,12 @@ public class BrandController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteBrand(@PathVariable Long id) {
-        brandService.deleteBrand(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        Optional<Brand> existingBrand = brandService.getBrandById(id);
+        if(existingBrand.isPresent()){
+            brandService.deleteBrand(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>( HttpStatus.NOT_FOUND);
+        }
     }
 }
