+++ src/main/java/com/java/demo/ecommerceapi/controller/CategoryController.java
@@ -8,9 +8,12 @@ import org.springframework.web.bind.annotation.*;
 import com.java.demo.ecommerceapi.model.Category;
 import com.java.demo.ecommerceapi.service.CategoryService;
 
+import io.swagger.v3.oas.annotations.tags.Tag;
+
 import java.util.List;
 import java.util.Optional;
 
+@Tag(name = "Authentication", description = "Endpoints for category management")
 @RestController
 @RequestMapping("/api/categories")
 public class CategoryController {
@@ -28,10 +31,10 @@ public class CategoryController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Optional<Category>> getCategoryById(@PathVariable Long id) {
+    public ResponseEntity<Category> getCategoryById(@PathVariable Long id) {
         Optional<Category> category = categoryService.getCategoryById(id);
         if (category.isPresent()) {
-            return new ResponseEntity<>(category, HttpStatus.OK);
+            return new ResponseEntity<>(category.get(), HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -40,7 +43,11 @@ public class CategoryController {
     @PostMapping
     public ResponseEntity<Category> createCategory(@RequestBody Category category) {
         Category createdCategory = categoryService.createCategory(category);
-        return new ResponseEntity<>(createdCategory, HttpStatus.CREATED);
+        if(createdCategory !=null){
+            return new ResponseEntity<>(createdCategory, HttpStatus.CREATED);
+        }else{
+            return new ResponseEntity<>( HttpStatus.BAD_REQUEST);
+        }
     }
 
     @PutMapping("/{id}")
@@ -55,7 +62,12 @@ public class CategoryController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCategory(@PathVariable Long id) {
-        categoryService.deleteCategory(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        Optional<Category> existinCategory = categoryService.getCategoryById(id);
+        if(existinCategory.isPresent()){
+            categoryService.deleteCategory(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>( HttpStatus.NOT_FOUND);
+        }
     }
 }
